Fix editable-install test crash and sensor return type

+++ custom_components/sma_meter/sensor.py
@@ -67,4 +67,4 @@ class SmaSensor(SmaEntity, SensorEntity):
         if scaled == int(scaled) and self.entity_description.scale != 1.0:
             return int(scaled)
 
-        return round(scaled, 3)
+        return float(round(scaled, 3))
